% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAlpha0.R
\name{calcAlpha0}
\alias{calcAlpha0}
\title{Calculate the needed value of alpha_0_outcome or alpha_0_treat}
\usage{
calcAlpha0(desired_prop, cov_dist, tol = 1e-04)
}
\arguments{
\item{desired_prop}{The desired event rate}

\item{cov_dist}{The covariate distribution. An integer 1-5 or one of the following:
'Ind Norm', 'Corr Norm', '5 Bern', '9 Bern', '10 Bern'}

\item{tol}{The maximum difference between the upper and lower estimate}
}
\value{
A numeric to be used as alpha_0_treat or alpha_0_outcome.
}
\description{
The form of the treatment and outcome regressions require values of
alpha_0_outcome or alpha_0_treat. In order to achieve a desired outcome
or treatment rate with a given covariate distribution, this value must
be calculated. It is not analytically solvable and thus it is approximated
using a simulation and bisection method.
}
