% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBeta.R
\name{calcBeta}
\alias{calcBeta}
\title{Calculate the needed value of beta to achieve a specified risk difference}
\usage{
calcBeta(true_risk_diff, cov_dist, alpha_0_outcome, tol = 1e-04)
}
\arguments{
\item{true_risk_diff}{The desired treated risk minus control risk}

\item{cov_dist}{The covariate distribution. An integer 1-5 or one of the following:
'Ind Norm', 'Corr Norm', '5 Bern', '9 Bern', '10 Bern'}

\item{alpha_0_outcome}{The calculated value alpha_0_outcome, obtained from calcAlpha0()}

\item{tol}{The maximum difference between the upper and lower estimate}
}
\value{
The value of beta that corresponds to the desired risk difference for
the given covariate distribution.
}
\description{
For the specified outcome regression, a value of beta must be specified as the
log-odds ratio relating the treatment to the outcome. This beta will have a
consequence on the true risk difference and thus beta must be calculated accordingly.
It is not analytically tractable and thus we estimate it using simulation and the
bisection method.
}
