% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMatchedRiskDiff.R
\name{calcMatchedRiskDiff}
\alias{calcMatchedRiskDiff}
\title{Conduct propensity score matching with a caliper and calculate the treatment effect}
\usage{
calcMatchedRiskDiff(data, gamma)
}
\arguments{
\item{data}{A data frame containing at least the columns:
'X1', ..., 'X10', 'Z'}

\item{gamma}{The caliper width in terms of pooled standard deviations}
}
\value{
The risk difference of the matched treated and control units.
}
\description{
The logit of the propensity scores are used to match control units to treated
units that are within the specified caliper. The risk difference (risk in treated
group minus risk in control group) is then calculated on the results matched sample.
}
