% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate a data frame containing 10 covariates, a treatment indicator, and a binary outcome}
\usage{
generateData(cov_dist, alpha_0_treat, alpha_0_outcome, beta)
}
\arguments{
\item{cov_dist}{The covariate distribution. An integer 1-5 or one of the following:
'Ind Norm', 'Corr Norm', '5 Bern', '9 Bern', '10 Bern'}

\item{alpha_0_treat}{The calculated value of alpha_0_treat, obtained from calcAlpha0()}

\item{alpha_0_outcome}{The calculated value alpha_0_outcome, obtained from calcAlpha0()}

\item{beta}{The calculated value of beta, obtained from calcBeta()}
}
\value{
A data frame with columns for the covariates, treatment indicator, and binary outcome,
as well as the logit and regular probabilities of treatment and outcome that were used
to generate them.
}
\description{
This function generates the needed data frame for the simulations,
which includes 10 covariates of a specified distribution, a treatment indicator,
and a binary outcome variable.
}
