% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Utility function to check if a number is a whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{The number to check}

\item{tol}{The tolerance}
}
\value{
Boolean stating whether x is a whole number
}
\description{
Utility function to check if a number is a whole number
}
