% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMOptimalCalipers.R
\docType{package}
\name{PSMOptimalCalipers}
\alias{PSMOptimalCalipers}
\title{PSMOptimalCalipers: Reproduces and Extends Simulations on Optimal Caliper Width in Propensity-Score Matching}
\description{
This package reproduces and extends figure 2 from the 2011 paper
paper 'Optimal caliper widths for propensity-score matching when estimating differences
in means and differences in proportions in observational studies' by Peter C. Austin.
Figure 2 examines the MSE of the predicted risk differences when using 50 different
propensity-score caliper widths from 0.05 to 2.5. The simulations all have 25 percent
treated patients and 75 percent control patients. We will extend these simulations by
varying the percentage of treated patients.
}
