% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulations.R
\name{runSimulations}
\alias{runSimulations}
\title{Run simulations for various covariate distributions, risk differences, and caliper widths}
\usage{
runSimulations(
  seed,
  cores,
  cov_dists,
  true_risk_diffs,
  gammas,
  N,
  p_treat,
  base_risk,
  file = NULL
)
}
\arguments{
\item{seed}{The seed to use for the random number generator}

\item{cores}{The number of cores to use}

\item{cov_dists}{A vector of covariate distributions for which to run simulations}

\item{true_risk_diffs}{A vector of true risk differences for which to run simulations}

\item{gammas}{A vector of caliper widths in terms of pooled standard deviations for which to run simulations}

\item{N}{The number of data sets to generate within the simulation}

\item{p_treat}{A numeric between 0 and 1 giving the proportion of units that should be treated (constant across simulations)}

\item{base_risk}{A numeric between 0 and 1 giving the base risk of the outcome if all units untreated}

\item{file}{Optional file to save workspace to after each set of N simulations}
}
\value{
A list with an entry for each covariate entry. That entry is a list with
an entry for each of the true risk differences. This entry is a matrix, with one
column for each gamma, which contains the set of simulation results:
the gamma value used, MSE of the matched estimator,
the bias reduction of using the matched estimator, the average estimate from the
full data, the average estimate from the matched data, the average time to go through
one data set, and the variance of the times to go through one data set.
}
\description{
This function runs simulations for a combination of covariate distributions,
risk differences and caliper widths and a given treatment proportion and base outcome risk.
The function can be parallelized with multiple cores, and reproduced with a seed value.
}
